requirements.txt
Flask==2.0.1
line-bot-sdk==1.19.0
discord.py==1.7.2
redis==3.5.3